Rework parser and implement ++ and -- operators.
The parser now keeps track of spaces within recurred statements.  Before it
simply would remove all spaces, turning statements like 6 + ++x into 6+++x
which couldn't get parsed by the parser for ++ and --.  Now the strings are
preserved with their spaces, and are simply stripped each time before going
back into the JIT.

Other than that, ++ and -- are now implemted for both prefix and postfix.
These simply translate internally to +=1 and -=1 when run, so ++x would be
translated to x+=1 and rerun through the JIT.

NOTE: ++ and -- CANNOT be used in math statements, so no '6 + ++x'.  This is a
limitation of the math parser, which will have to be corrected in the future.